+++ .github/workflows/clean.yaml
@@ -1,0 +1,21 @@
+name: Clean files
+
+on:
+  push:
+    branches:
+      - main
+
+jobs:
+  clean:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v4
+
+      - name: Run script
+        run: node ./.github/clean.js
+
+      - name: Commit changed files
+        uses: stefanzweifel/git-auto-commit-action@v5
+        with:
+          commit_message: "Clean files"
